Attempted to fix bugs with the timer.
Bug 1: the timer sometimes resets to zero. Commented out a line that resets it to zero, (but there is another line later in the code that may also lead to the bug).

Bug 2: the timer switches off when I click on a OPEN IN NEW TAB link because of tab updating. Added a check to see if the tab is active